projects/capstone/starter/myPyboProject/myPyboBackend/requirements.txt
alembic==1.10.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ecdsa==0.18.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pyasn1==0.4.8
python-jose==3.3.0
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.6
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
